[fs] sysfs: Do not warn about missing kernfs_node if kobj is not active
Message-id: <20180126140821.7647-2-vgoyal@redhat.com>
Patchwork-id: 204413
O-Subject: [RHEL7.5 PATCH 1/1] sysfs: Do not warn about missing kernfs_node if kobj is not active
Bugzilla: 1534568
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1534568

upstream commit: RHEL only

After backport of usernamespace patches, QE reported that sysfs started
displaying warnings once in a while when an HBA is removed or rmmod is done.

This warning basically says that sysfs_remove_group() is looking for a
kernfs_node in the rb-tree of parent node and it can't find it. So it
warns about it.

This primarily has happened due to the way kernfs does the node
removal/cleanup. Now a directory/node removal is recursive. That is if
top directory goes away then all the children node will be drained
and disconnected (__kernfs_remove()).

When kernfs does this, it drains all the children node first. (kernfs_drain()).
In that call it calls ->release() function of object and I think owner of
object should cleanup all object related state. That way there are no
conflicts between recursive removal disconnecting the object and owner trying
to clean it up later and be surprised.

Ofcourse for this to work, bunch of changes were done upstream in drivers,
scsi and common stack etc. But they all seemed to be scattered and its not
known how many such fixes went in. And it might even require a backport of
enhancements in kworker area. So this works fine upstream but does not work
fine in RHEL.

So once in a while what happens is that kernfs has done the hiearchy
cleanup and driver cleanup code is probably still executing in some kworker
thread and by the time that code tries to do cleanup, it can't find the
node (it has already been cleaned up) and it warns about it.

While going through __kernfs_remove(), I noticed that kernfs marks all
children node inactive first. And then it gets into the business of
going through all the children and clean these up. So this patch checks
if parent kobject is still active or not. If it is not active, that means
kernfs is already cleaning it up and don't warn. If parent kobject is
still active, then display the warning.

This is still not perfect and there could still be some cases where we do
not display the warning when we should have. But I think that's not a huge
issue. And reason being that this warning as of now does not break any
functionality. It is just warning that kobject cleanup is not happening
in correct order. Figure it out and fix it.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>