[x86] spec_ctrl: make ipbp_enabled read-only
Message-id: <71cd3fb869233b75fabafd3a5e01af0ec5692864.1516996615.git.jpoimboe@redhat.com>
Patchwork-id: 204435
O-Subject: [RHEL7.5 PATCH 05/47] x86/spec_ctrl: make ipbp_enabled read-only
Bugzilla: 1535644
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>

Now that we don't have ibpb_enabled=2, it no longer makes sense to have
the variable be user-controlled.  It only adds to the complexity of the
user experience, as well as to the code.  Instead we can make
ibpb_enabled be automatically selected based on the ibrs_enabled
setting.

For similar reasons, make the 'noibpb' cmdline option a no-op.

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>