[x86] spec_ctrl: detect unretpolined modules
Message-id: <6d5249ff83eafbc823fa6fa36fe8808b2a2e7972.1516996615.git.jpoimboe@redhat.com>
Patchwork-id: 204459
O-Subject: [RHEL7.5 PATCH 28/47] x86/spec_ctrl: detect unretpolined modules
Bugzilla: 1535644
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>

With retpolines as the Spectre v2 mitigation, loading a module that was
compiled without retpolines can potentially make the kernel vulnerable.

In such a case, we could in theory disable retpolines and switch the
system to IBRS, but that might be too drastic.

Instead, just print a warning and downgrade the reported mitigation to
"Vulnerable".

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>