[x86] spec_ctrl: add ibrs_enabled=3 (ibrs_user)
Message-id: <1a311d6b83037fcc458f3cd539f589eab9e6fb6d.1516996615.git.jpoimboe@redhat.com>
Patchwork-id: 204464
O-Subject: [RHEL7.5 PATCH 30/47] x86/spec_ctrl: add ibrs_enabled=3 (ibrs_user)
Bugzilla: 1535644
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>

When running with retpolines, there's a Spectre variant 2 attack vector
from the guest to qemu (through the host kernel).  The mitigation is
either to recompile qemu (and all its libraries) with retpolines; or to
enable IBRS in user space (and disable it in kernel space for
performance reasons).

Add support for user space IBRS with a new user-space only
ibrs_enabled=3 option.  It can be combined with retpolines to protect
the host user space from the guest, as well as to protect user space
from user-user SMT-based attacks.

Rename the existing 'ibrs_user' option to the more accurate
'ibrs_always', and then call the new option 'ibrs_user'.  The naming is
internal to the code so it's not user-visible interface change.

Suggested-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>