[x86] unwind: fix livepatch regression with CALL_NOSPEC macro
Message-id: <ff900dfa1d99e3382e216e1b91af5cb970b7a4f5.1516996615.git.jpoimboe@redhat.com>
Patchwork-id: 204470
O-Subject: [RHEL7.5 PATCH 37/47] x86/unwind: fix livepatch regression with CALL_NOSPEC macro
Bugzilla: 1535644
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>

Livepatch is broken with the new CALL_NOSPEC macro when retpolines are
disabled (or when the AMD retpoline is used).

Livepatch relies on the unwinder's use of the UNWIND_END_OF_STACK macro,
which requires the CALL instruction to be immediately before it.
Unfortunately the new CALL_NOSPEC macro doesn't always do that.

Fix it by adding labels around the CALL_NOSPEC in ret_from_fork.

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>