[x86] spec_ctrl: Avoid returns in IBRS-disabled regions
Message-id: <213389dfece8c46cce22f371e4d7687bbd038354.1516996615.git.jpoimboe@redhat.com>
Patchwork-id: 204473
O-Subject: [RHEL7.5 PATCH 39/47] x86/spec_ctrl: Avoid returns in IBRS-disabled regions
Bugzilla: 1535644
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Jon Masters <jcm@redhat.com>

On Skylake, if the RSB becomes empty (underflow), and IBRS is not on,
the next 'ret' uses the BTB to speculate its return target, which is
bad.

This means that the IBRS-disabled regions are exposed.  If we get an IRQ
or NMI which has a long enough call stack to fill up the RSB, then when
it returns, the RSB will be empty, and any returns in the IBRS-disabled
regions will be speculated from the BTB.

The quick and dirty fix is to ensure that all function calls in non-IBRS
paths are always inline, so that we never get a 'ret' until IBRS is
re-enabled.

Another possible way to fix it is to stuff RSB on IRQ and NMI exit for
Skylake, but that's more difficult and this is good enough for now.

Suggested-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>