[netdrv] iwlwifi: mvm: fix security bug in PN checking
Message-id: <20180207094759.GB13981@redhat.com>
Patchwork-id: 205059
O-Subject: [PATCH RHEL7.5 v2] iwlwifi: mvm: fix security bug in PN checking
Bugzilla: 1538028
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Michal Schmidt <mschmidt@redhat.com>

BZ#1538028

Upstream:
Patch not submitted by Intel yet. It's placed on backport-iwlwifi.git tree:
https://git.kernel.org/pub/scm/linux/kernel/git/iwlwifi/backport-iwlwifi.git/commit/?id=efd36048c666fde5baebf79f774a70f8d5a4ad16

Description:
A previous patch allowed the same PN for packets originating from the
same AMSDU by copying PN only for the last packet in the series.

This however is bogus since we cannot assume the last frame will be
received on the same queue, and if it is received on a different ueue
we will end up not incrementing the PN and possibly let the next
packet to have the same PN and pass through.

Change the logic instead to driver explicitly indicate for the second
sub frame and on to be allowed to have the same PN as the first
subframe. Indicate it to mac80211 as well for the fallback queue.

Brew:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15121835

Test:
Tested by Karl Hastings, patch fixes performance issue on 5GHz band.
---

v2: add PATCH prefix to the topic.
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>