[tools] perf: Fix swap for samples with raw data
Message-id: <20180211215941.GB5821@krava>
Patchwork-id: 205611
O-Subject: [PATCHv2 RHEL7.5 BZ1458228] perf tools: Fix swap for samples with raw data
Bugzilla: 1458228
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <brdeoliv@redhat.com>

Bugzilla: 1458228
https://bugzilla.redhat.com/show_bug.cgi?id=1458228

upstream
========
commit f9d8adb345d7adbb2d3431eea73beb89c8d6d612
Author: Jiri Olsa <jolsa@kernel.org>
Date:   Wed Nov 29 19:43:46 2017 +0100

description
===========
When we detect a different endianity we swap event before processing.
It's tricky for samples because we have no idea what's inside. We treat
it as an array of u64s, swap them and later on we swap back parts which
are different.

We mangle this way also the tracepoint raw data, which ends up in report
showing wrong data:

  1.95%  comm=Q^B pid=29285 prio=16777216 target_cpu=000
  1.67%  comm=l^B pid=0 prio=16777216 target_cpu=000

Luckily the traceevent library handles the endianity by itself (thank
you Steven!), so we can pass the RAW data directly in the other
endianity.

  2.51%  comm=beah-rhts-task pid=1175 prio=120 target_cpu=002
  2.23%  comm=kworker/0:0 pid=11566 prio=120 target_cpu=000

The fix is basically to swap back the raw data if different endianity is
detected.

[ Add util/memswap.c to python-ext-sources to link missing mem_bswap_64() ]

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>