[netdrv] ibmvnic: Fix rx queue cleanup for non-fatal resets
Message-id: <1518897451-3700-4-git-send-email-drosario@redhat.com>
Patchwork-id: 205792
O-Subject: [PATCH 3/7] ibmvnic: Fix rx queue cleanup for non-fatal resets
Bugzilla: 1544356
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1544356

DESCRIPTION:
------------
At some point, a check was added to exit the polling routine during resets.
This makes sense for most reset conditions, but for a non-fatal error, we
expect the polling routine to continue running to properly clean up the rx
queues. This patch checks if we are performing a non-fatal reset and if we
are, continues normal polling operation.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 3468656fd7599b0cb1092bb1ee717d1a984e93ee

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=3468656fd7599b0cb1092bb1ee717d1a984e93ee
----------------
Desnes Augusto Nunes do Rosario

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>