[netdrv] ibmvnic: Reset long term map ID counter
Message-id: <1518897451-3700-7-git-send-email-drosario@redhat.com>
Patchwork-id: 205795
O-Subject: [PATCH 6/7] ibmvnic: Reset long term map ID counter
Bugzilla: 1544356
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1544356

DESCRIPTION:
------------
When allocating RX or TX buffer pools, the driver needs to provide a
unique mapping ID to firmware for each pool. This value is assigned
using a counter which is incremented after a new pool is created. The
ID can be an integer ranging from 1-255. When migrating to a device
that requests a different number of queues, this value was not being
reset properly. As a result, after enough migrations, the counter
exceeded the upper bound and pool creation failed. This is fixed by
resetting the counter to one in this case.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit faefaa97215a0c05105d7ae180fe1a3b5979ad1f

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=faefaa97215a0c05105d7ae180fe1a3b5979ad1f
----------------
Desnes Augusto Nunes do Rosario

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>