[tools] perf intel-pt: Improve sample timestamp
Message-id: <20180219120557.28358-47-jolsa@redhat.com>
Patchwork-id: 205855
O-Subject: [PATCH RHEL7.6 BZ1532861 46/98] perf intel-pt: Improve sample timestamp
Bugzilla: 1532861
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1532861
https://bugzilla.redhat.com/show_bug.cgi?id=1532861

upstream
========
commit 3f04d98e972b59706bd43d6cc75efac91f8fba50
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri May 26 11:17:03 2017 +0300

description
===========
The decoder uses its current timestamp in samples. Usually that is a
timestamp that has already passed, but in some cases it is a timestamp
for a branch that the decoder is walking towards, and consequently
hasn't reached. Improve that situation by using the pkt_state to
determine when to use the current or previous timestamp.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>