[tools] perf intel-pt: Use FUP always when scanning for an IP
Message-id: <20180219120557.28358-52-jolsa@redhat.com>
Patchwork-id: 205842
O-Subject: [PATCH RHEL7.6 BZ1532861 51/98] perf intel-pt: Use FUP always when scanning for an IP
Bugzilla: 1532861
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1532861
https://bugzilla.redhat.com/show_bug.cgi?id=1532861

upstream
========
commit 622b7a47b843c78626f40c1d1aeef8483383fba2
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri May 26 11:17:08 2017 +0300

description
===========
The decoder will try to use branch packets to find an IP to start decoding
or to recover from errors. Currently the FUP packet is used only in the
case of an overflow, however there is no reason for that to be a special
case. So just use FUP always when scanning for an IP.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>