[tools] perf intel-pt: Clear FUP flag on error
Message-id: <20180219120557.28358-53-jolsa@redhat.com>
Patchwork-id: 205853
O-Subject: [PATCH RHEL7.6 BZ1532861 52/98] perf intel-pt: Clear FUP flag on error
Bugzilla: 1532861
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1532861
https://bugzilla.redhat.com/show_bug.cgi?id=1532861

upstream
========
commit 6a558f12dbe85437acbdec5e149ea07b5554eced
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri May 26 11:17:09 2017 +0300

description
===========
Sometimes a FUP packet is associated with a TSX transaction and a flag is
set to indicate that. Ensure that flag is cleared on any error condition
because at that point the decoder can no longer assume it is correct.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>