[tools] perf intel-pt: Synthesize new power and "ptwrite" events
Message-id: <20180219120557.28358-89-jolsa@redhat.com>
Patchwork-id: 205905
O-Subject: [PATCH RHEL7.6 BZ1532861 88/98] perf intel-pt: Synthesize new power and "ptwrite" events
Bugzilla: 1532861
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1532861
https://bugzilla.redhat.com/show_bug.cgi?id=1532861

upstream
========
commit 3797307576191d7fb4c974cd461188162ac36f33
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Jun 30 11:36:45 2017 +0300

description
===========
Synthesize new power and ptwrite events.

Power events report changes to C-state but I have also added support
for the existing CBR (core-to-bus ratio) packet and included that
when outputting power events.

The PTWRITE packet is associated with the new "ptwrite" instruction,
which is essentially just a way to stuff a 32 or 64 bit value into the
PT trace.

More details can be found in the patches that add documentation and in
the Intel SDM.

[ Copy the description of such packet from the patchkit cover message ]

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>