[mm] fs: rework do_invalidatepage
Message-id: <c55c4911-8b96-691a-a3d5-88b9051aa8e3@redhat.com>
Patchwork-id: 205982
O-Subject: [PATCH RHEL7] fs: rework do_invalidatepage
Bugzilla: 1546079
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Bill O'Donnell <billodo@redhat.com>

Bug: 1546079[xfstests xfs/259 regression] Bug at mm/truncate.c:332
Build: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15382483
Testing: local with xfstests & gluster
	KT1: https://beaker.engineering.redhat.com/jobs/2322956
Upstream: No, RHEL-specific

This is a RHEL only patch, which fixes a subtle KABI breakage as well
as a crash in xfstests and in gluster when fallocate is used.

In commit ee8da60267192 ("[fs] mm: teach truncate_inode_pages_range()
to handle non page aligned ranges"), truncate_inode_pages_range was
modified to work with non-page-aligned ends, but only for file systems
that implemented a newly-added, kabi-extended ->invalidate_range()
address space operation.  A BUG_ON was added to catch instances where
truncate_inode_pages_range() was called with a mis-aligned end when the
underlying file system did not support that via the updated aop:

    void truncate_inode_pages_range(struct address_space *mapping,
                                    loff_t lstart, loff_t lend)
    {
            ...
            if (!inode_has_invalidate_range(mapping->host))
                    BUG_ON(partial_end);

This was initially sufficient because no core code called into
truncate_inode_pages_range() with an unaligned end, and filesystems
that did could implement ->invalidate_range() to handle it as needed.

Later patches added callers of this function which could pass in non-
aligned ends, such as the most recent patch from Ming, commit
4ea65f63ac94e ("[block] Invalidate cache on discard v2").  Fuse
was affected as well, because its fallocate/punch hole path directly
called truncate_inode_pages_range for the punched-out region, but it
did not define any ->invalidate[_range]() aop.  But hitting the
BUG_ON in that case was incorrect, because address spaces with no
form of ->invalidate (like fuse) fall back to the generic
block_invalidatepage_range which actually /can/ handle it.

This patch (from Brian Foster, really) manages all this within
do_invalidatepage_range().  It safely tests whether ->invalidatepage_range
exists, and uses it if so.  If no ->invalidatepage variant exists,
it uses the unaligned-capable block_invalidatepage_range function.
If the older ->invalidatepage aop is implemented, it still will
BUG_ON an unaligned endpoint, which ->invalidatepage cannot handle
in RHEL7, but page-aligned ends are allowed to proceed.

That last BUG_ON should be safe for now at least because there is no
core kernel code which should be passing in an unaligned end for anything
but a block device; for the two cases we've seen, the discard ioctl
is acting on the block device address space which will use
block_invalidatepage_range, and the fallocate/punch case was actually
caused by fuse itself requesting the unaligned range, not the core
punch code.  If any filesystem wants to call truncate on partial
pages, it will need to implement ->invalidatepage_range that can
handle it.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
---

Big thanks to Jeff & Brian & Bill & Carlos & Lukas and everyone else
for chiming in on this problem.

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>