[x86] spec_ctrl: avoid rmb() on full retpoline kernels
Message-id: <20180227213819.6692-23-pbonzini@redhat.com>
Patchwork-id: 206091
O-Subject: [RHEL7.5 PATCH 22/23] RHEL: spec_ctrl: avoid rmb() on full retpoline kernels
Bugzilla: 1537379
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

If the kernel is "full retpoline", there are no indirect branches within
the kernel and it is running with IBRS=0.  In that case, we don't risk
incorrectly speculating an indirect branch after vmexit or kernel entry,
and the expensive rmb() can be removed.

Saves another ~50 cycles on KVM vmexit.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>