[redhat] x86/makefile: minor adjust to GCC retpoline error message
Message-id: <d94a5508bad08c0791722dcef75658ad54dc67fc.1519838435.git.aquini@redhat.com>
Patchwork-id: 206131
O-Subject: [RHEL7 PATCH] [redhat] arch/x86/makefile: minor adjust to GCC retpoline error message
Bugzilla: 1550235
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

RHEL-only patch, not upstream, minor nit to a build error message.

It was brought to our attention that the word "Toolchain" in the error
message spitted by the compiler in case of it not having support
to retpolines is too broad and, apparently, is becoming source of
confusion.

This patch disambiguates that message by switching Toolchain by Compiler.

Signed-off-by: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>