[mm] revert kvmalloc: stress the vmalloc path in the debugging kernel
Message-id: <x49muzs3lb9.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 206134
O-Subject: [RHEL7.5 PATCH] Revert "[mm] kvmalloc: stress the vmalloc path in the debugging kernel"
Bugzilla: 1550094
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <brdeoliv@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

BZ: 1550094
Brew: 15449330

This reverts commit ea376cc55bc3faf358de464c1111fd967c2c0f09.

It looks like the patch to find invalid uses of vmalloc memory was a
success!  Unfortunately, we merged it at the last minute, and now we're
seeing test failures (panics) with the -debug kernel.  These failures
haven't been seen in the wild yet (on a non-debug kernel), so let's
defer root-causing and fixing them until the beginning of the 7.6
development cycle.

I'll file a separate bug to re-introduce this patch for 7.6.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>

---
Note: I'm still waiting on testing feedback to ensure this does fix the
issues reported in the bug.  I kicked off kt1 tests, and Eric Sandeen is
running xfstests, but progess is slow.

Signed-off-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>