[netdrv] ibmvnic: Free and re-allocate scrqs when tx/rx scrqs change
Message-id: <1520867260-11758-6-git-send-email-drosario@redhat.com>
Patchwork-id: 206480
O-Subject: [RHEL-7.6 BZ1547983 PATCH 05/10] ibmvnic: Free and re-allocate scrqs when tx/rx scrqs change
Bugzilla: 1547983
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1547983

DESCRIPTION:
------------
When the driver resets it is possible that the number of tx/rx
sub-crqs can change. This patch handles this so that the driver does
not try to access non-existent sub-crqs.

The count for releasing sub crqs depends on the adapter state. The
active queue count is not set in probe, so if we are relasing in probe
state we use the request queue count.

Additionally, a parameter is added to release_sub_crqs() so that
we know if the h_call to free the sub-crq needs to be made. In
the reset path we have to do a reset of the main crq, which is
a free followed by a register of the main crq. The free of main
crq results in all of the sub crq's being free'ed. When updating
sub-crq count in the reset path we do not want to h_free the
sub-crqs, they are already free'ed.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit d7c0ef36bde03af0d119d895bba4cc0362e79dbb

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=d7c0ef36bde03af0d119d895bba4cc0362e79dbb
----------------
Desnes Augusto Nunes do Rosario

PROPOSED PATCH:

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>