[redhat] spec: properly clean up bootloader menuentry in case new-kernel-pkg fails to create the initramfs file
Message-id: <56903f1e309a99965da8b32ea0f3743a7c349699.1521485670.git.aquini@redhat.com>
Patchwork-id: 206982
O-Subject: [RHEL7 RHBZ#1422262 PATCH v2] redhat: spec: properly clean up bootloader menuentry in case new-kernel-pkg fails to create the initramfs file
Bugzilla: 1422262
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <brdeoliv@redhat.com>
RH-Acked-by: Rodrigo Freire <rfreire@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

In order to sort out yet another one nightmarish RPM install dependency issue,
changes for RHEL6 bug 528063 de-coupled steps that get the initramfs created
and introduce the bootloader menuentry for the newly installed kernel thus
unwittingly opening a race window we're closing with this patch.

As initramfs files need to be compiled after the kernel is installed on hardware
and their size will depend on the particularities of the system at hand, there's
no simple way to determine the size of that file when building the RPM pre-shipping.
It happens that, due to the size constraints usually observed, sometimes /boot partition
can run out of space to accomodate the iniramfs file. In such cases, when installing
a new kernel via RPM, one can get the system bootloader config in a inconsistent state
that will not allow the system to re-boot properly. This issue happens because
the, now de-coupled, steps that introduce a bootloader menu-entry for the new kernel
and create its respective initramfs file are happening out of order.

It's important to state that changes introduced for RHEL6 bug 528063 are not wrong,
although they created the space for this little annoyance at hand, thus we cannot
afford just reverting them. Instead, this patch goes the simplest way possible,
and whithout much churn we just cope with the corner case, if/when it happens,
catching up the exception and backing out the recently introduced menuentry
so the system can reboot fine on the old-kernel. This fix is important for
customers that do automatize the cycle update/reboot, preventing them from
rendering systems halted due to missing-initramfs-panic at boot time.

Signed-off-by: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>