[netdrv] ibmvnic: Report queue stops and restarts as debug output
Message-id: <1521563038-2118-5-git-send-email-drosario@redhat.com>
Patchwork-id: 207047
O-Subject: [RHEL-7.6 BZ1552020 PATCH 4/5] ibmvnic: Report queue stops and restarts as debug output
Bugzilla: 1552020
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1552020

DESCRIPTION:
------------
It's not necessary to report each time a queue is stopped and restarted
as an informational message. Change that to be a debug message so that
it can be observed if needed but not printed by default.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 0aecb13ce3abe1372c7306c7d1b86ff748296abb

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=0aecb13ce3abe1372c7306c7d1b86ff748296abb
----------------
Desnes A. Nunes do Rosario

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>