[netdrv] ibmvnic: Do not attempt to login if RX or TX queues are not allocated
Message-id: <1521563038-2118-6-git-send-email-drosario@redhat.com>
Patchwork-id: 207049
O-Subject: [RHEL-7.6 BZ1552020 PATCH 5/5] ibmvnic: Do not attempt to login if RX or TX queues are not allocated
Bugzilla: 1552020
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1552020

DESCRIPTION:
------------
If a device reset fails for some reason, TX and RX queue resources
could be released. If a user attempts to open the device in this scenario,
it may result in a kernel panic as the driver tries to access this
memory. To fix this, include a check before device login that TX/RX
queues are still there before enabling the device. In addition, return a
value that can be checked in case of any errors to avoid waiting for a
completion that will never come.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 20a8ab744ff799ccedd35aba0d3139782f341bed

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=20a8ab744ff799ccedd35aba0d3139782f341bed
----------------
Desnes A. Nunes do Rosario

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>