[netdrv] vmxnet3: avoid xmit reset due to a race in vmxnet3
Message-id: <20180402121814.32246-2-nhorman@redhat.com>
Patchwork-id: 207698
O-Subject: [RHEL7.6 PATCH 1/2] vmxnet3: avoid xmit reset due to a race in vmxnet3
Bugzilla: 1558685
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

From: Neil Horman <nhorman@redhat.com>

The field txNumDeferred is used by the driver to keep track of the number
of packets it has pushed to the emulation. The driver increments it on
pushing the packet to the emulation and the emulation resets it to 0 at
the end of the transmit.

There is a possibility of a race either when (a) ESX is under heavy load or
(b) workload inside VM is of low packet rate.

This race results in xmit hangs when network coalescing is disabled. This
change creates a local copy of txNumDeferred and uses it to perform ring
arithmetic.

Reported-by: Noriho Tanaka <ntanaka@vmware.com>
Signed-off-by: Ronak Doshi <doshir@vmware.com>
Acked-by: Shrikrishna Khare <skhare@vmware.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7a4c003d6921e2af215f4790aa43a292bdc78be0)

BZ: 1558685
Test: Tested by vmware
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=15676759
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>