[netdrv] ibmvnic: Fix reset return from closed state
Message-id: <1522704904-30716-10-git-send-email-drosario@redhat.com>
Patchwork-id: 207797
O-Subject: [RHEL-7.6 BZ1559635 PATCH 09/19] ibmvnic: Fix reset return from closed state
Bugzilla: 1559635
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1559635

DESCRIPTION:
------------
The case in which we handle a reset from the state where the device is
closed seems to be bugged for all types of reset. For most types of reset
we currently exit the reset routine correctly, but don't set the state to
indicate that we are back in the "closed" state. For some specific cases,
we don't exit the reset routine at all and resetting will cause a closed
device to be opened.

This patch fixes the problem by unconditionally checking the reset_state
and correctly setting the adapter state before returning.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit e676d81c8990f511d60698a1a8abaa438b3f9d3d

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=e676d81c8990f511d60698a1a8abaa438b3f9d3d
----------------
Desnes A. Nunes do Rosario

PROPOSED PATCH:

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>