[netdrv] ibmvnic: Disable irqs before exiting reset from closed state
Message-id: <1522704904-30716-20-git-send-email-drosario@redhat.com>
Patchwork-id: 207806
O-Subject: [RHEL-7.6 BZ1559635 PATCH 19/19] ibmvnic: Disable irqs before exiting reset from closed state
Bugzilla: 1559635
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1559635

DESCRIPTION:
------------
When the driver is closed, all the associated irqs are disabled. In the
event that the driver exits a reset in the closed state, we should be
consistent with the state we are in directly after a close. So before we
exit the reset routine, all irqs should be disabled as well. This will
prevent the irqs from being enabled twice in this case and reporting a
number of noisy warning traces.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 6095e590d277ecec54805c4999335406f08855fe

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net-next.git/commit/?id=6095e590d277ecec54805c4999335406f08855fe
----------------
Desnes A. Nunes do Rosario

PROPOSED PATCH:

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>