[nvme] nvmet-fc: correct ref counting error when deferred rcv used
Message-id: <1523042961-48197-90-git-send-email-dmilburn@redhat.com>
Patchwork-id: 208711
O-Subject: [RHEL7.6 PATCH BZ 1519689 89/96] nvmet-fc: correct ref counting error when deferred rcv used
Bugzilla: 1519689
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: James Smart <jsmart2021@gmail.com>

Whenever a cmd is received a reference is taken while looking up the
queue. The reference is removed after the cmd is done as the iod is
returned for reuse. The fod may be reused for a deferred (recevied but
no job context) cmd.  Existing code removes the reference only if the
fod is not reused for another command. Given the fod may be used for
one or more ios, although a reference was taken per io, it won't be
matched on the frees.

Remove the reference on every fod free. This pairs the references to
each io.

Signed-off-by: James Smart <james.smart@broadcom.com>
Reviewed-by: Sagi Grimberg <sagi@grimberg.me>
Signed-off-by: Christoph Hellwig <hch@lst.de>
(cherry picked from commit 619c62dcc62b957d17cccde2081cad527b020883)
Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>