[netdrv] ibmvnic: Handle all login error conditions
Message-id: <1523914843-17715-2-git-send-email-drosario@redhat.com>
Patchwork-id: 209563
O-Subject: [RHEL-7.6 BZ1566205 PATCH 1/2] ibmvnic: Handle all login error conditions
Bugzilla: 1566205
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

From: "Desnes A. Nunes do Rosario" <drosario@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1566205

DESCRIPTION:
------------
There is a bug in handling the possible return codes from sending the
login CRQ. The current code treats any non-success return value,
minus failure to send the crq and a timeout waiting for a login response,
as a need to re-send the login CRQ. This can put the drive in an
infinite loop of trying to login when getting return values other
that a partial success such as a return code of aborted. For these
scenarios the login will not ever succeed at this point and the
driver would need to be reset again.

To resolve this loop trying to login is updated to only retry the
login if the driver gets a return code of a partial success. Other
return codes are treated as an error and the driver returns an error
from ibmvnic_login().

To avoid infinite looping in the partial success return cases, the
number of retries is capped at the maximum number of supported
queues. This value was chosen because the driver does a renegotiation
of capabilities which sets the number of queues possible and allows
the driver to attempt a login for possible value for the number
of queues supported.

Signed-off-by: Nathan Fontenot <nfont@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

UPSTREAM STATUS:
----------------
commit 64d92aa2c9fe490ceffc440d7648ce369cd6cc3c

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=64d92aa2c9fe490ceffc440d7648ce369cd6cc3c
----------------
Desnes A. Nunes do Rosario

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>