[tools] perf annotate: Implement visual marker for macro fusion
Message-id: <20180419185703.28105-21-jolsa@redhat.com>
Patchwork-id: 209944
O-Subject: [PATCH RHEL7.6 BZ1540862 020/245] perf annotate: Implement visual marker for macro fusion
Bugzilla: 1540862
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1540862
https://bugzilla.redhat.com/show_bug.cgi?id=1540862

upstream
========
commit 7e63a13a266da652f82731b845b5c35dd866ec7e
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Jul 7 13:06:35 2017 +0800

description
===========
For marking fused instructions clearly this patch adds a line before the
first instruction of pair and joins it with the arrow of the jump to its
target.

For example, when "je" is selected in annotate view, the line before
cmpl is displayed and joins the arrow of "je".

          cmpl   $0x0,argp_program_version_hook
 81.93    je     20
            lock   cmpxchg %esi,0x38a9a4(%rip)
           jne    29
           jmp    43
 11.47 20:cmpxch %esi,0x38a999(%rip)

That means the cmpl+je is a fused instruction pair and they should be
considered together.

Changelog:

v3: Use Arnaldo's fix to improve the arrow origin rendering.  To get the
    evsel->evlist->env->cpuid, save the evsel in annotate_browser.

v2: new function "ins__is_fused" to check if the instructions are fused.
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>