[tools] perf util: Add const modifier to buf in "writen" function
Message-id: <20180419185703.28105-64-jolsa@redhat.com>
Patchwork-id: 210009
O-Subject: [PATCH RHEL7.6 BZ1540862 063/245] perf util: Add const modifier to buf in "writen" function
Bugzilla: 1540862
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1540862
https://bugzilla.redhat.com/show_bug.cgi?id=1540862

upstream
========
commit 7c72440506e2108494bab3b68a4118fa61a9dbf4
Author: David Carrillo-Cisneros <davidcc@google.com>
Date: Mon Jul 17 21:25:37 2017 -0700

description
===========
Make buf in helper function "writen" constant to simplify the life of
its callers.

This requires to hack a cast of buf prior to passing it to "ion" which
is simpler than the alternative of reworking the "ion" function to
provide a read and a write paths, the latter with constant buf argument.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>