[tools] perf trace: Use the syscall_fmt formatters without a tracepoint
Message-id: <20180419185703.28105-87-jolsa@redhat.com>
Patchwork-id: 210166
O-Subject: [PATCH RHEL7.6 BZ1540862 086/245] perf trace: Use the syscall_fmt formatters without a tracepoint
Bugzilla: 1540862
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1540862
https://bugzilla.redhat.com/show_bug.cgi?id=1540862

upstream
========
commit d032d79e2dcb56e13678bf2cc7b36957ef827c32
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Jul 19 14:36:13 2017 -0300

description
===========
Previously we only used the syscall_fmt when we had sc->tp_format set,
i.e. when we found the (enter, exit) pair in tracefs/events/syscalls/.

But we really only need to use what is in sc->arg_fmt to apply the arg
beautifiers to the syscall argument values, so do it.

With this we will be able to provide formatters to the "clone" syscall,
which doesn't have entries in tracefs/events/syscalls/.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>