[tools] perf test: Add infrastructure to run shell based tests
Message-id: <20180419185703.28105-156-jolsa@redhat.com>
Patchwork-id: 210184
O-Subject: [PATCH RHEL7.6 BZ1540862 155/245] perf test: Add infrastructure to run shell based tests
Bugzilla: 1540862
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1540862
https://bugzilla.redhat.com/show_bug.cgi?id=1540862

upstream
========
commit 1209b273a25ee60a267b606bea77e068b8556a8d
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Aug 3 16:49:15 2017 -0300

description
===========
To allow testing by directly using perf tools in scripts, checking that
the effects on the system are the ones expected and that the output
produced is as well the desired one.

For instance, adding a probe at a well known location with 'perf probe',
then checking that the results from using that probe to record are the
desired ones, etc.

The next csets will introduce tests using this new testing
infrastructure.

The scripts should return 0 for Ok, 1 for FAIL and 2 for SKIP.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>