[tools] perf srcline: Do not consider empty files as valid srclines
Message-id: <20180419185703.28105-164-jolsa@redhat.com>
Patchwork-id: 210117
O-Subject: [PATCH RHEL7.6 BZ1540862 163/245] perf srcline: Do not consider empty files as valid srclines
Bugzilla: 1540862
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1540862
https://bugzilla.redhat.com/show_bug.cgi?id=1540862

upstream
========
commit d964b1cdbd94f359f1f65f81440be84ceb45978e
Author: Milian Wolff <milian.wolff@kdab.com>
Date: Sun Aug 6 23:24:45 2017 +0200

description
===========
Sometimes we get a non-null, but empty, string for the filename from
bfd. This then results in srclines of the form ":0", which is different
from the canonical SRCLINE_UNKNOWN in the form "??:0".  Set the file to
NULL if it is empty to fix this.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>